requirements.txt
# Web framework
fastapi>=0.109
uvicorn[standard]>=0.27

# HTTP client
httpx[http2]>=0.27

# Browser automation
playwright>=1.40

# HTML parsing
beautifulsoup4>=4.12
lxml>=5.0

# Search
ddgs>=9.0

# Database
supabase>=2.3

# Utilities
tqdm>=4.66
python-dotenv>=1.0
